Get branch name from substitutions field
Problem

A cloud build's build message has the union field `revision` which can
be one of the following.

- `branchName`
- `tagName`
- `commitSha`

Originally we used `branchName` to get the branch name of a build but
cloud build suddenly prioritizes commitSha. As a result we get empty
branch name.

Solution

Use `substitutions.BRANCH_NAME` to get branch name.